+++ Scripts/SPSWakeUP.ps1
@@ -4,15 +4,10 @@
 
     .DESCRIPTION
     SPSWakeUp is a PowerShell script tool to warm up all site collection in your SharePoint environment.
-    It's compatible with all supported versions for SharePoint (2010 to 2019).
+    It's compatible with all supported versions for SharePoint (2016 to Subscription Edition).
     Use WebRequest object in multi-thread to download JS, CSS and Pictures files,
     Log script results in log file,
-    Email nofications,
     Configure automatically prerequisites for a best warm-up,
-
-    .PARAMETER InputFile
-    Need parameter input file, example:
-    PS D:\> E:\SCRIPT\SPSWakeUP.ps1 -InputFile 'E:\SCRIPT\SPSWakeUP.psd1'
 
     .PARAMETER Install
     Use the switch Install parameter if you want to add the warmup script in taskscheduler
@@ -32,17 +27,22 @@
     and only warmup the root web of the site collection.
     PS D:\> E:\SCRIPT\SPSWakeUP.ps1 -OnlyRootWeb
 
+    .PARAMETER Transcript
+    Use the boolean Transcript parameter if you want to start Transcrit PowerShell Feature.
+    PS D:\> E:\SCRIPT\SPSWakeUP.ps1 -Transcript:$True
+
     .EXAMPLE
-    SPSWakeUP.ps1 -InputFile 'E:\SCRIPT\SPSWakeUP.psd1'
     SPSWakeUP.ps1 -Install -InstallAccount (Get-Credential)
     SPSWakeUP.ps1 -Uninstall
     SPSWakeUP.ps1 -OnlyRootWeb
+    SPSWakeUP.ps1 -Transcript:$True
 
     .NOTES
     FileName:	SPSWakeUP.ps1
-    Author:		luigilink (Jean-Cyril DROUHIN)
-    Date:		July 16, 2021
-    Version:	2.6.1
+    Authors:	luigilink (Jean-Cyril DROUHIN)
+                Nutsoft (Des Finkenzeller)
+    Date:		May 05, 2023
+    Version:	2.7.0
     Licence:	MIT License
 
     .LINK
@@ -51,10 +51,6 @@
 #>
 param
 (
-    [Parameter(Position = 0)]
-    [System.String]
-    $InputFile,
-
     [Parameter(Position = 1)]
     [switch]
     $Install,
@@ -69,53 +65,44 @@ param
 
     [Parameter(Position = 4)]
     [switch]
-    $OnlyRootWeb
+    $OnlyRootWeb,
+
+    [Parameter(Position = 5)]
+    [System.Boolean]
+    $Transcript = $false
 )
 
 Clear-Host
 $Host.UI.RawUI.WindowTitle = "WarmUP script running on $env:COMPUTERNAME"
 
 # Define variable
-$spsWakeupVersion = '2.6.1'
-$currentUser = ([Security.Principal.WindowsIdentity]::GetCurrent()).Name
-$scriptRootPath = Split-Path -parent $MyInvocation.MyCommand.Definition
+$spsWakeupVersion = '2.7.0'
+$currentUser      = ([Security.Principal.WindowsIdentity]::GetCurrent()).Name
+$scriptRootPath   = Split-Path -parent $MyInvocation.MyCommand.Definition
+$hostEntries      =  New-Object -TypeName System.Collections.Generic.List[string]
+$hostsFile        = "$env:windir\System32\drivers\etc\HOSTS"
+$hostsFileCopy    = $hostsFile + '.' + (Get-Date -UFormat "%y%m%d%H%M%S").ToString() + '.copy'
 
-$pathLogFile = Join-Path -Path $scriptRootPath -ChildPath ('SPSWakeUP_script_' + (Get-Date -Format yyyy-MM-dd_H-mm) + '.log')
-$logFileContent =  New-Object -TypeName System.Collections.Generic.List[string]
-
-$hostEntries =  New-Object -TypeName System.Collections.Generic.List[string]
-$hostsFile = "$env:windir\System32\drivers\etc\HOSTS"
-$hostsFileCopy = $hostsFile + '.' + (Get-Date -UFormat "%y%m%d%H%M%S").ToString() + '.copy'
-
-# Get the content of the SPSWakeUP.xml file
-if (-not($InputFile))
-{
-    $InputFile = Join-Path -Path $scriptRootPath -ChildPath 'SPSWakeUP.psd1'
-}
-if (Test-Path -Path $InputFile)
-{
-    $dataConfig = Import-LocalizedData -BaseDirectory $scriptRootPath -FileName 'SPSWakeUP.psd1'
+#Start Transcript parameter is equal to True
+if ($Transcript) {
+    $pathLogFile  = Join-Path -Path $scriptRootPath -ChildPath ('SPSWakeUP_script_' + (Get-Date -Format yyyy-MM-dd_H-mm) + '.log')
+    Start-Transcript -Path $pathLogFile -IncludeInvocationHeader
 }
 
 #Check UserName and Password if Install parameter is used
-if ($Install)
-{
-    if ($null -eq $InstallAccount)
-    {
+if ($Install) {
+    if ($null -eq $InstallAccount) {
         Write-Warning -Message ('SPSWakeUp: Install parameter is set. Please set also InstallAccount ' + `
-                                "parameter. `nSee https://spwakeup.com for details.")
+                "parameter. `nSee https://spwakeup.com for details.")
         Break
     }
-    else
-    {
-        $UserName = $InstallAccount.UserName
-        $Password = $InstallAccount.GetNetworkCredential().Password
-
+    else {
+        $UserName      = $InstallAccount.UserName
+        $Password      = $InstallAccount.GetNetworkCredential().Password
         $currentDomain = 'LDAP://' + ([ADSI]'').distinguishedName
         Write-Output "Checking Account `"$UserName`" ..."
-        $dom = New-Object System.DirectoryServices.DirectoryEntry($currentDomain,$UserName,$Password)
-        if ($null -eq $dom.Path)
-        {
+        $dom = New-Object System.DirectoryServices.DirectoryEntry($currentDomain, $UserName, $Password)
+        if ($null -eq $dom.Path) {
             Write-Warning -Message "Password Invalid for user:`"$UserName`""
             Break
         }
@@ -126,28 +113,6 @@ if ($Install)
 # INTERNAL FUNCTIONS
 # ====================================================================================
 #region logging and trap exception
-<#
-    .SYNOPSIS
-    Displays a standardized verbose message.
-
-    .PARAMETER Message
-    String containing the key of the localized verbose message.
-#>
-function Write-VerboseMessage
-{
-    [CmdletBinding()]
-    [Alias()]
-    [OutputType([String])]
-    param
-    (
-        [Parameter(Mandatory=$true)]
-        [ValidateNotNull()]
-        [System.String]
-        $Message
-    )
-
-    Write-Verbose -Message ((Get-Date -format yyyy-MM-dd_HH-mm-ss) + ": $Message");
-}
 # ===================================================================================
 # Func: Write-LogException
 # Desc: write Exception in powershell session and in error file
@@ -170,87 +135,12 @@ function Write-LogException
         $Message
     )
 
+    $pathErrLog   = Join-Path -Path $scriptRootPath -ChildPath (((Get-Date).Ticks.ToString()) + '_errlog.xml')
+    $messageEvent = $Message.Exception.Message + "`r`n" + 'For more informations, see errlog.xml file:' + $pathErrLog
+
     Write-Warning -Message $Message.Exception.Message
-    $pathErrLog = Join-Path -Path $scriptRootPath -ChildPath (((Get-Date).Ticks.ToString()) + '_errlog.xml')
     Export-Clixml -Path $pathErrLog -InputObject $Message -Depth 3
-    Write-LogContent -Message 'For more informations, see errlog.xml file:'
-    Write-LogContent -Message $pathErrLog
-}
-# ===================================================================================
-# Func: Save-LogFile
-# Desc: Save the log file in current folder
-# ===================================================================================
-function Save-LogFile
-{
-    param
-    (
-        [Parameter(Mandatory=$true)]
-        [System.String]
-        $Path
-    )
-
-    $pathLogFile = New-Object -TypeName System.IO.StreamWriter($Path)
-    foreach ($logFileC in $logFileContent)
-    {
-        $pathLogFile.WriteLine($logFileC)
-    }
-
-    $pathLogFile.Close()
-}
-# ===================================================================================
-# Func: Write-LogContent
-# Desc: Add Content in log file and write-output in console
-# ===================================================================================
-function Write-LogContent
-{
-    [CmdletBinding()]
-    [Alias()]
-    [OutputType([String])]
-    param
-    (
-        [Parameter(Mandatory=$true)]
-        [ValidateNotNull()]
-        [System.String]
-        $Message
-    )
-
-    Write-Output $Message
-
-    $logFileContent.Add($Message)
-}
-# ===================================================================================
-# Func: Send-SPSLog
-# Desc: Send Email with log file in attachment
-# ===================================================================================
-function Send-SPSLog
-{
-    param
-    (
-        [Parameter(Mandatory=$true)]
-        $MailAttachment,
-
-        [Parameter(Mandatory=$true)]
-        $MailBody
-    )
-
-    if ($dataConfig.Settings.EmailNotification.Enable)
-    {
-        $mailAddress = $dataConfig.Settings.EmailNotification.EmailAddress
-        $smtpServer = $dataConfig.Settings.EmailNotification.SMTPServer
-        $mailSubject = "Automated Script - SPSWakeUP Urls - $env:COMPUTERNAME"
-
-        Write-LogContent -Message '--------------------------------------------------------------'
-        Write-LogContent -Message "Sending Email with Log file to $mailAddress ..."
-        try
-        {
-            Send-MailMessage -To $mailAddress -From $mailAddress -Subject $mailSubject -Body $MailBody -BodyAsHtml -SmtpServer $smtpServer -Attachments $MailAttachment -ea stop
-            Write-LogContent -Message "Email sent successfully to $mailAddress"
-        }
-        catch
-        {
-            Write-LogException -Message $_
-        }
-    }
+    Write-EventLog -LogName Application -Source 'SPSWakeUp' -EntryType Warning -EventId 7701 -Message $messageEvent
 }
 # ===================================================================================
 # Func: Clear-SPSLog
@@ -261,43 +151,34 @@ function Clear-SPSLog
     param
     (
         [Parameter(Mandatory=$true)]
-        [System.String]$path
+        [ValidateNotNullOrEmpty()]
+        [System.String]
+        $path,
+
+        [Parameter()]
+        [System.UInt32]
+        $days = 30
     )
 
-    if (Test-Path $path)
-    {
-        # Days of logs that will be remaining after log cleanup.
-        $days = $dataConfig.Settings.CleanLogsDays
-
+    if (Test-Path $path) {
         # Get the current date
         $Now = Get-Date
-
         # Definie the extension of log files
         $Extension = '*.log'
-
         # Define LastWriteTime parameter based on $days
         $LastWrite = $Now.AddDays(-$days)
-
         # Get files based on lastwrite filter and specified folder
         $files = Get-Childitem -Path "$path\*.*" -Include $Extension | Where-Object -FilterScript {
             $_.LastWriteTime -le "$LastWrite"
         }
 
-        if ($files)
-        {
-            Write-LogContent -Message '--------------------------------------------------------------'
-            Write-LogContent -Message "Cleaning log files in $path ..."
-            foreach ($file in $files)
-            {
-                if ($null -ne $file)
-                {
-                    Write-LogContent -Message "Deleting file $file ..."
-                    Remove-Item $file.FullName | out-null
-                }
-                else
-                {
-                    Write-LogContent -Message 'No more log files to delete'
-                    Write-LogContent -Message '--------------------------------------------------------------'
+        if ($files) {
+            Write-Output '--------------------------------------------------------------'
+            Write-Output "Cleaning log files in $path ..."
+            foreach ($file in $files) {
+                if ($null -ne $file) {
+                    Write-Output "Deleting file $file ..."
+                    Remove-Item $file.FullName | Out-Null
                 }
             }
         }
@@ -335,24 +216,21 @@ function Add-SPSTask
         $_.Name -eq $TaskName
     }
     $TaskCmd = 'C:\Windows\System32\WindowsPowerShell\v1.0\powershell.exe'
-    $inputFileFullPath = (Get-Item $InputFile).FullName;
     $TaskCmdArg =
 @"
--Command Start-Process "$PSHOME\powershell.exe" -Verb RunAs -ArgumentList "'-ExecutionPolicy Bypass ""$path\SPSWakeUP.ps1 -inputFile $inputFileFullPath""'"
+-Command Start-Process "$PSHOME\powershell.exe" -Verb RunAs -ArgumentList "'-ExecutionPolicy Bypass ""$path\SPSWakeUP.ps1""'"
 "@
 
-    if ($TaskSPSWKP)
-    {
+    if ($TaskSPSWKP) {
         Write-Warning -Message 'Shedule Task already exists - skipping.'
     }
-    else
-    {
-        Write-LogContent -Message '--------------------------------------------------------------'
-        Write-LogContent -Message 'Adding SPSWakeUP script in Task Scheduler Service ...'
+    else {
+        Write-Output '--------------------------------------------------------------'
+        Write-Output 'Adding SPSWakeUP script in Task Scheduler Service ...'
 
         # Get Credentials for Task Schedule
         $TaskAuthor = ([Security.Principal.WindowsIdentity]::GetCurrent()).Name
-        $TaskUser =  $UserName
+        $TaskUser = $UserName
         $TaskUserPwd = $Password
 
         # Add a New Task Schedule
@@ -393,13 +271,11 @@ function Add-SPSTask
         $TaskAction = $TaskSchd.Actions.Create(0)
         $TaskAction.Path = $TaskCmd
         $TaskAction.Arguments = $TaskCmdArg
-        try
-        {
+        try {
             $TaskFolder.RegisterTaskDefinition( $TaskName, $TaskSchd, 6, $TaskUser , $TaskUserPwd , 1)
-            Write-LogContent -Message 'Successfully added SPSWakeUP script in Task Scheduler Service'
+            Write-Output 'Successfully added SPSWakeUP script in Task Scheduler Service'
         }
-        catch
-        {
+        catch {
             Write-LogException -Message $_
         }
     }
@@ -420,67 +296,36 @@ function Remove-SPSTask
     $TaskSPSWKP = $TaskFolder.GetTasks(0) | Where-Object -FilterScript {
         $_.Name -eq $TaskName
     }
-
-    if ($null -eq $TaskSPSWKP)
-    {
+    if ($null -eq $TaskSPSWKP) {
         Write-Warning -Message 'Shedule Task already removed - skipping.'
     }
-    else
-    {
-        Write-LogContent -Message '--------------------------------------------------------------'
-        Write-LogContent -Message 'Removing SPSWakeUP script in Task Scheduler Service ...'
-
-        try
-        {
-            $TaskFolder.DeleteTask($TaskName,$null)
-            Write-LogContent -Message 'Successfully removed SPSWakeUP script from Task Scheduler Service'
+    else {
+        Write-Output '--------------------------------------------------------------'
+        Write-Output 'Removing SPSWakeUP script in Task Scheduler Service ...'
+        try {
+            $TaskFolder.DeleteTask($TaskName, $null)
+            Write-Output 'Successfully removed SPSWakeUP script from Task Scheduler Service'
         }
-        catch
-        {
+        catch {
             Write-LogException -Message $_
         }
     }
 }
 #endregion
 
-#region Load SharePoint Powershell Snapin for SharePoint 2010, 2013 & 2016
+#region Load SharePoint Powershell Snapin for SharePoint Server
 # ===================================================================================
 # Name: 		Add-PSSharePoint
 # Description:	Load SharePoint Powershell Snapin
 # ===================================================================================
 function Add-PSSharePoint
 {
-    if ($null -eq (Get-PsSnapin | Where-Object -FilterScript {$_.Name -eq 'Microsoft.SharePoint.PowerShell'}))
-    {
-        Write-LogContent -Message '--------------------------------------------------------------'
-        Write-LogContent -Message 'Loading SharePoint Powershell Snapin ...'
-        Add-PsSnapin Microsoft.SharePoint.PowerShell -ErrorAction Stop | Out-Null
-        Write-LogContent -Message '--------------------------------------------------------------'
+    if ($null -eq (Get-PSSnapin | Where-Object -FilterScript { $_.Name -eq 'Microsoft.SharePoint.PowerShell' })) {
+        Write-Output '--------------------------------------------------------------'
+        Write-Output 'Loading SharePoint Powershell Snapin ...'
+        Add-PSSnapin Microsoft.SharePoint.PowerShell -ErrorAction Stop | Out-Null
+        Write-Output '--------------------------------------------------------------'
     }
-}
-# ===================================================================================
-# Name: 		Add-RASharePoint
-# Description:	Load SharePoint Assembly for SharePoint 2007, 2010, 2013 & 2016
-# ===================================================================================
-function Add-RASharePoint
-{
-    Write-LogContent -Message '--------------------------------------------------------------'
-    Write-LogContent -Message 'Loading Microsoft.SharePoint Assembly ...'
-    [System.Reflection.Assembly]::LoadWithPartialName('Microsoft.SharePoint') | Out-Null
-    Write-LogContent -Message '--------------------------------------------------------------'
-}
-# ===================================================================================
-# Name: 		Add-SystemWeb
-# Description:	Load System.Web with Reflection Assembly
-# ===================================================================================
-function Add-SystemWeb
-{
-
-    Write-LogContent -Message '--------------------------------------------------------------'
-    Write-LogContent -Message 'Loading System.Web ...'
-    [System.Reflection.Assembly]::LoadWithPartialName('system.web') | Out-Null
-    Write-LogContent -Message '--------------------------------------------------------------'
-
 }
 # ===================================================================================
 # Name: 		Get-SPSThrottleLimit
@@ -489,21 +334,17 @@ function Add-SystemWeb
 function Get-SPSThrottleLimit
 {
     [int]$NumThrottle = 8
-
     # Get Number Of Throttle Limit
-    try
-    {
+    try {
         $cimInstanceProc = @(Get-CimInstance -ClassName Win32_Processor)
         $cimInstanceSocket = $cimInstanceProc.count
         $numLogicalCpu = $cimInstanceProc[0].NumberOfLogicalProcessors * $cimInstanceSocket
 
         $NumThrottle = @{ $true = 10; $false = 2 * $numLogicalCpu }[$numLogicalCpu -ge 8]
     }
-    catch
-    {
+    catch {
         Write-Warning -Message $_
     }
-
     return $NumThrottle
 }
 #endregion
@@ -517,17 +358,14 @@ function Get-SPSVersion
 {
     # location in registry to get info about installed software
     $regLoc = Get-ChildItem HKLM:\Software\Microsoft\Windows\CurrentVersion\Uninstall
-
     # Get SharePoint Products and language packs
     $programs = $regLoc |  Where-Object -FilterScript {
         $_.PsPath -like '*\Office*'
     } | ForEach-Object -Process { Get-ItemProperty $_.PsPath }
-
     # output the info about Products and Language Packs
     $spsVersion = $programs | Where-Object -FilterScript {
         $_.DisplayName -like '*SharePoint Server*'
     }
-
     # Return SharePoint version
     $spsVersion.DisplayVersion
 }
@@ -594,39 +432,34 @@ function Get-SPSSitesUrl
         $topologySvcUrl = 'http://localhost:32843/Topology/topology.svc'
         [void]$tbSitesURL.Add((Add-SPSSitesUrl -Url $topologySvcUrl))
 
-        # Get url of CentralAdmin if include in input xml file
-        if ($dataConfig.Settings.IncludeCentralAdmin -eq $true)
-        {
-            $webAppADM = Get-SPWebApplication -IncludeCentralAdministration | Where-Object -FilterScript {
-                $_.IsAdministrationWebApplication
-            }
-            $siteADM = $webAppADM.Url
-            # Most useful administration pages
-            [void]$tbSitesURL.Add((Add-SPSSitesUrl -Url $siteADM))
-            [void]$tbSitesURL.Add((Add-SPSSitesUrl -Url $siteADM'Lists/HealthReports/AllItems.aspx'))
-            [void]$tbSitesURL.Add((Add-SPSSitesUrl -Url $siteADM'_admin/FarmServers.aspx'))
-            [void]$tbSitesURL.Add((Add-SPSSitesUrl -Url $siteADM'_admin/Server.aspx'))
-            [void]$tbSitesURL.Add((Add-SPSSitesUrl -Url $siteADM'_admin/WebApplicationList.aspx'))
-            [void]$tbSitesURL.Add((Add-SPSSitesUrl -Url $siteADM'_admin/ServiceApplications.aspx'))
+        $webAppADM = Get-SPWebApplication -IncludeCentralAdministration | Where-Object -FilterScript {
+            $_.IsAdministrationWebApplication
+        }
+        $siteADM = $webAppADM.Url
+        # Most useful administration pages
+        [void]$tbSitesURL.Add((Add-SPSSitesUrl -Url $siteADM))
+        [void]$tbSitesURL.Add((Add-SPSSitesUrl -Url $siteADM'Lists/HealthReports/AllItems.aspx'))
+        [void]$tbSitesURL.Add((Add-SPSSitesUrl -Url $siteADM'_admin/FarmServers.aspx'))
+        [void]$tbSitesURL.Add((Add-SPSSitesUrl -Url $siteADM'_admin/Server.aspx'))
+        [void]$tbSitesURL.Add((Add-SPSSitesUrl -Url $siteADM'_admin/WebApplicationList.aspx'))
+        [void]$tbSitesURL.Add((Add-SPSSitesUrl -Url $siteADM'_admin/ServiceApplications.aspx'))
 
-            # Quick launch top links
-            [void]$tbSitesURL.Add((Add-SPSSitesUrl -Url $siteADM'applications.aspx'))
-            [void]$tbSitesURL.Add((Add-SPSSitesUrl -Url $siteADM'systemsettings.aspx'))
-            [void]$tbSitesURL.Add((Add-SPSSitesUrl -Url $siteADM'monitoring.aspx'))
-            [void]$tbSitesURL.Add((Add-SPSSitesUrl -Url $siteADM'backups.aspx'))
-            [void]$tbSitesURL.Add((Add-SPSSitesUrl -Url $siteADM'security.aspx'))
-            [void]$tbSitesURL.Add((Add-SPSSitesUrl -Url $siteADM'upgradeandmigration.aspx'))
-            [void]$tbSitesURL.Add((Add-SPSSitesUrl -Url $siteADM'apps.aspx'))
-            [void]$tbSitesURL.Add((Add-SPSSitesUrl -Url $siteADM'generalapplicationsettings.aspx'))
+        # Quick launch top links
+        [void]$tbSitesURL.Add((Add-SPSSitesUrl -Url $siteADM'applications.aspx'))
+        [void]$tbSitesURL.Add((Add-SPSSitesUrl -Url $siteADM'systemsettings.aspx'))
+        [void]$tbSitesURL.Add((Add-SPSSitesUrl -Url $siteADM'monitoring.aspx'))
+        [void]$tbSitesURL.Add((Add-SPSSitesUrl -Url $siteADM'backups.aspx'))
+        [void]$tbSitesURL.Add((Add-SPSSitesUrl -Url $siteADM'security.aspx'))
+        [void]$tbSitesURL.Add((Add-SPSSitesUrl -Url $siteADM'upgradeandmigration.aspx'))
+        [void]$tbSitesURL.Add((Add-SPSSitesUrl -Url $siteADM'apps.aspx'))
+        [void]$tbSitesURL.Add((Add-SPSSitesUrl -Url $siteADM'generalapplicationsettings.aspx'))
 
-            # Get Service Application Urls
-            $sa = Get-SPServiceApplication
-            $linkUrls = $sa | ForEach-Object {$_.ManageLink.Url} | Select-Object -Unique
-            foreach ($linkUrl in $linkUrls)
-            {
-                $siteADMSA = $linkUrl.TrimStart('/')
-                [void]$tbSitesURL.Add((Add-SPSSitesUrl -Url "$siteADM$siteADMSA"))
-            }
+        # Get Service Application Urls
+        $sa = Get-SPServiceApplication
+        $linkUrls = $sa | ForEach-Object {$_.ManageLink.Url} | Select-Object -Unique
+        foreach ($linkUrl in $linkUrls) {
+            $siteADMSA = $linkUrl.TrimStart('/')
+            [void]$tbSitesURL.Add((Add-SPSSitesUrl -Url "$siteADM$siteADMSA"))
         }
 
         # Get Url of all site collection
@@ -729,17 +562,15 @@ function Get-SPSHSNCUrl
         $_.HostHeaderIsSiteName -eq $true
     }
 
-    foreach ($HSNC in $HSNCs)
-    {
-        if ($HSNC.Url -notmatch 'sitemaster-')
-        {
+    foreach ($HSNC in $HSNCs) {
+        if ($HSNC.Url -notmatch 'sitemaster-') {
             [void]$hsncURL.Add($HSNC.Url)
             Add-SPSHostEntry -Url $HSNC.Url
         }
         $HSNC.Dispose()
     }
 
-    $hsncURL
+    return $hsncURL
 }
 # ===================================================================================
 # Name: 		Get-SPSWebAppUrl
@@ -753,19 +584,17 @@ function Get-SPSWebAppUrl
     $webAppURL = New-Object -TypeName System.Collections.ArrayList
     $webApps = Get-SPWebApplication
 
-    foreach ($webapp in $webApps)
-    {
+    foreach ($webapp in $webApps) {
         [void]$webAppURL.Add($webapp.GetResponseUri('Default').AbsoluteUri)
         $spSrvIsInUri = Get-SPServer | Where-Object -FilterScript {
             $webapp.GetResponseUri('Default').AbsoluteUri -match $_.Name
         }
-        if ($null -eq $spSrvIsInUri)
-        {
+        if ($null -eq $spSrvIsInUri) {
             Add-SPSHostEntry -Url $webapp.GetResponseUri('Default').AbsoluteUri
         }
     }
 
-    $webAppURL
+    return $webAppURL
 }
 #endregion
 
@@ -901,7 +730,7 @@ function Invoke-SPSWebRequest
     }
     catch
     {
-        Write-LogContent -Message 'An error occurred invoking multi-threading function'
+        Write-Output 'An error occurred invoking multi-threading function'
         Write-LogException -Message $_
     }
 
@@ -916,103 +745,18 @@ function Invoke-SPSWebRequest
 #region Configuration and permission
 # ===================================================================================
 # Func: Disable-LoopbackCheck
-# Desc: Disable Loopback Check
+# Desc: This setting usually kicks out a 401 error when you try to navigate to sites
+#       that resolve to a loopback address e.g.  127.0.0.1
 # ===================================================================================
-function Disable-LoopbackCheck
-{
-    param
-    (
-        [Parameter(Mandatory=$true)]$hostNameList
-    )
-
-    # Disable the Loopback Check on stand alone demo servers.
-    # This setting usually kicks out a 401 error when you try to navigate to sites that resolve to a loopback address e.g.  127.0.0.1
-    if ($dataConfig.Settings.DisableLoopbackCheck -eq $true)
-    {
-
-        $lsaPath = 'HKLM:\System\CurrentControlSet\Control\Lsa'
-        $lsaPathValue = Get-ItemProperty -path $lsaPath
-        if (-not ($lsaPathValue.DisableLoopbackCheck -eq '1'))
-        {
-            Write-LogContent -Message 'Disabling Loopback Check...'
-            New-ItemProperty HKLM:\System\CurrentControlSet\Control\Lsa -Name 'DisableLoopbackCheck' -value '1' -PropertyType dword -Force | Out-Null
-        }
-        else
-        {
-            Write-LogContent -Message 'Loopback Check already Disabled - skipping.'
-        }
+function Disable-LoopbackCheck {
+    $lsaPath      = 'HKLM:\System\CurrentControlSet\Control\Lsa'
+    $lsaPathValue = Get-ItemProperty -path $lsaPath
+    if (-not ($lsaPathValue.DisableLoopbackCheck -eq '1')) {
+        Write-Output 'Disabling Loopback Check...'
+        New-ItemProperty -Path HKLM:\System\CurrentControlSet\Control\Lsa -Name 'DisableLoopbackCheck' -value '1' -PropertyType dword -Force | Out-Null
     }
-    else
-    {
-        $lsaPath = 'HKLM:\System\CurrentControlSet\Control\Lsa'
-        $paramPath = 'HKLM:System\CurrentControlSet\Services\LanmanServer\Parameters'
-        $mvaPath = 'HKLM:\System\CurrentControlSet\Control\Lsa\MSV1_0'
-        $lsaPathValue = Get-ItemProperty -path $lsaPath
-        $paramPathValue = Get-ItemProperty -path $paramPath
-
-        if ($lsaPathValue.DisableLoopbackCheck -eq '1')
-        {
-            Write-LogContent -Message 'Disabling Loopback Check - Back to default value ...'
-            New-ItemProperty $lsaPath -Name 'DisableLoopbackCheck' -value '0' -PropertyType dword -Force | Out-Null
-        }
-
-        if (-not($paramPathValue.DisableStrictNameChecking -eq '1'))
-        {
-            Write-LogContent -Message 'Disabling Strict Name Checking ...'
-            New-ItemProperty $paramPath -Name 'DisableStrictNameChecking' -value '1' -PropertyType dword -Force | Out-Null
-        }
-
-        $BackCoName = Get-ItemProperty -Path $mvaPath -Name BackConnectionHostNames -ea SilentlyContinue
-        if (!($BackCoName))
-        {
-            New-ItemProperty $mvaPath -Name 'BackConnectionHostNames' -PropertyType multistring -Force | Out-Null
-        }
-        foreach ($hostName in $hostNameList)
-        {
-            if (!($BackCoName.BackConnectionHostNames -like "*$hostName*"))
-            {
-                Write-LogContent -Message "Add $hostName in BackConnectionHostNames regedit key ..."
-                $BackCoNameNew = $BackCoName.BackConnectionHostNames + "$hostName"
-                New-ItemProperty $mvaPath -Name 'BackConnectionHostNames' -Value $BackCoNameNew -PropertyType multistring -Force | Out-Null
-            }
-        }
-    }
-}
-# ====================================================================================
-# Func: Backup-HostsFile
-# Desc: Backup HOSTS File System
-# ====================================================================================
-function Backup-HostsFile
-{
-    Param
-    (
-        [Parameter(Mandatory=$true)]$hostsFilePath,
-        [Parameter(Mandatory=$true)]$hostsBackupPath
-    )
-
-    if ($dataConfig.Settings.AddURLsToHOSTS.Enable -eq $true)
-    {
-        Write-LogContent -Message "Backing up $hostsFilePath file to:"
-        Write-LogContent -Message "$hostsBackupPath"
-        Copy-Item $hostsFilePath -Destination $hostsBackupPath -Force
-    }
-}
-# ====================================================================================
-# Func: Restore-HostsFile
-# Desc: Restore previous HOSTS File System
-# ====================================================================================
-function Restore-HostsFile
-{
-    Param
-    (
-        [Parameter(Mandatory=$true)]$hostsFilePath,
-        [Parameter(Mandatory=$true)]$hostsBackupPath
-    )
-    if ($dataConfig.Settings.AddURLsToHOSTS.Enable -eq $true -AND $dataConfig.Settings.AddURLsToHOSTS.KeepOriginal -eq $true)
-    {
-        Write-LogContent -Message "Restoring $hostsBackupPath file to:"
-        Write-LogContent -Message "$hostsFilePath"
-        Copy-Item $hostsBackupPath -Destination $hostsFilePath -Force
+    else {
+        Write-Output 'Loopback Check already Disabled - skipping.'
     }
 }
 # ====================================================================================
@@ -1023,35 +767,30 @@ function Clear-HostsFileCopy
 {
     Param
     (
-        [Parameter(Mandatory=$true)]$hostsFilePath
+        [Parameter(Mandatory=$true)]
+        [ValidateNotNullOrEmpty()]
+        [System.String]
+        $hostsFilePath,
+
+        [Parameter()]
+        [System.UInt32]
+        $numberFiles = 10
     )
 
     $hostsFolderPath = Split-Path $hostsFilePath
-    if (Test-Path $hostsFolderPath)
-    {
-        # Number of files that will be remaining after backup cleanup.
-        $numberFiles = $dataConfig.Settings.AddURLsToHOSTS.Retention
+    if (Test-Path $hostsFolderPath) {
         # Definie the extension of log files
         $extension = '*.copy'
 
         # Get files with .copy extension, sort them by name, from most recent to oldest and skip the first numberFiles variable
-        $copyFiles = Get-Childitem -Path "$hostsFolderPath\*.*" -Include $extension | Sort-Object -Descending -Property Name | Select-Object -Skip $numberFiles
-
-        if ($copyFiles)
-        {
-            Write-LogContent -Message '--------------------------------------------------------------'
-            Write-LogContent -Message "Cleaning backup HOSTS files in $hostsFolderPath ..."
-            foreach ($copyFile in $copyFiles)
-            {
-                if ($null -ne $copyFile)
-                {
-                    Write-LogContent -Message "   * Deleting File $copyFile ..."
-                    Remove-Item $copyFile.FullName | out-null
-                }
-                Else
-                {
-                    Write-LogContent -Message 'No more backup HOSTS files to delete '
-                    Write-LogContent -Message '--------------------------------------------------------------'
+        $copyFiles = Get-ChildItem -Path "$hostsFolderPath\*.*" -Include $extension | Sort-Object -Descending -Property Name | Select-Object -Skip $numberFiles
+        if ($copyFiles) {
+            Write-Output '--------------------------------------------------------------'
+            Write-Output "Cleaning backup HOSTS files in $hostsFolderPath ..."
+            foreach ($copyFile in $copyFiles) {
+                if ($null -ne $copyFile) {
+                    Write-Output "   * Deleting File $copyFile ..."
+                    Remove-Item $copyFile.FullName | Out-Null
                 }
             }
         }
@@ -1065,26 +804,15 @@ function Add-HostsEntry
 {
     param
     (
-        [Parameter(Mandatory=$true)]$hostNameList
+        [Parameter(Mandatory=$true)]
+        [ValidateNotNullOrEmpty()]
+        [System.String[]]
+        $hostNameList
     )
 
-    if ($dataConfig.Settings.AddURLsToHOSTS.Enable -eq $true -and $hostNameList)
-    {
+    if ($null -ne $hostNameList) {
         $hostsContentFile =  New-Object System.Collections.Generic.List[string]
-        # Check if the IPv4Address configured in XML Input file is reachable
-        $hostIPV4Addr = $dataConfig.Settings.AddURLsToHOSTS.IPv4Address
-        Write-LogContent -Message "Testing connection (via Ping) to `"$hostIPV4Addr`"..."
-        $canConnect = Test-Connection $hostIPV4Addr -Count 1 -Quiet
-        if ($canConnect)
-        {
-            Write-LogContent -Message "IPv4Address $hostIPV4Addr will be used in HOSTS File during WarmUP ..."
-        }
-        else
-        {
-            Write-LogContent -Message '   * IPv4Address not valid in Input XML File, 127.0.0.1 will be used in HOSTS File'
-            $hostIPV4Addr = '127.0.0.1'
-        }
-
+        $hostIPV4Addr     = '127.0.0.1'
         $hostsContentFile.Add("
 # Copyright (c) 1993-2009 Microsoft Corp.
 #
@@ -1104,17 +832,13 @@ function Add-HostsEntry
 #      102.54.94.97     rhino.acme.com          # source server
 #       38.25.63.10     x.acme.com              # x client host
 ")
-
-        if ($dataConfig.Settings.AddURLsToHOSTS.ListRevocationUrl -eq $true){$hostsContentFile.Add("127.0.0.1 `t crl.microsoft.com")}
-        ForEach ($hostname in $hostNameList)
-        {
+        $hostsContentFile.Add("127.0.0.1 `t crl.microsoft.com")
+        foreach ($hostname in $hostNameList) {
             # Remove http or https information to keep only HostName or FQDN
-            if ($hostname.Contains(':'))
-            {
-                Write-LogContent -Message "$hostname cannot be added in HOSTS File, only web applications with 80 or 443 port are added."
+            if ($hostname.Contains(':')) {
+                Write-Warning -Message "$hostname cannot be added in HOSTS File, only web applications with 80 or 443 port are added."
             }
-            Else
-            {
+            else {
                 $hostsContentFile.Add("$hostIPV4Addr `t $hostname")
             }
         }
@@ -1140,104 +864,43 @@ function Add-SPSUserPolicy
         $UserName
     )
 
-    Write-LogContent -Message '--------------------------------------------------------------'
-    Write-LogContent -Message "Add Read Access to $UserName for All Web Applications ..."
-    foreach ($url in $Urls)
-    {
-        try
-        {
+    Write-Output '--------------------------------------------------------------'
+    Write-Output "Add Read Access to $UserName for All Web Applications ..."
+    foreach ($url in $Urls) {
+        try {
             $webapp = [Microsoft.SharePoint.Administration.SPWebApplication]::Lookup("$url")
             $displayName = 'SPSWakeUP Account'
 
             # If the web app is not Central Administration
-            if ($webapp.IsAdministrationWebApplication -eq $false)
-            {
+            if ($webapp.IsAdministrationWebApplication -eq $false) {
                 # If the web app is using Claims auth, change the user accounts to the proper syntax
-                if ($webapp.UseClaimsAuthentication -eq $true)
-                {
+                if ($webapp.UseClaimsAuthentication -eq $true) {
                     $user = (New-SPClaimsPrincipal -identity $UserName -identitytype 1).ToEncodedString()
                 }
-                else
-                {
+                else {
                     $user = $UserName
                 }
-                Write-LogContent -Message "Checking Read access for $user account to $url..."
+                Write-Output "Checking Read access for $user account to $url..."
                 [Microsoft.SharePoint.Administration.SPPolicyCollection]$policies = $webapp.Policies
                 $policyExist = $policies | Where-Object -FilterScript {
                     $_.Displayname -eq 'SPSWakeUP Account'
                 }
 
-                if (-not ($policyExist))
-                {
-                    Write-LogContent -Message "Applying Read access for $user account to $url..."
+                if (-not ($policyExist)) {
+                    Write-Output "Applying Read access for $user account to $url..."
                     [Microsoft.SharePoint.Administration.SPPolicy]$policy = $policies.Add($user, $displayName)
                     $policyRole = $webApp.PolicyRoles.GetSpecialRole([Microsoft.SharePoint.Administration.SPPolicyRoleType]::FullRead)
-                    if ($null -ne $policyRole)
-                    {
+                    if ($null -ne $policyRole) {
                         $policy.PolicyRoleBindings.Add($policyRole)
                     }
                     $webapp.Update()
-                    Write-LogContent -Message "Done Applying Read access for `"$user`" account to `"$url`""
+                    Write-Output "Done Applying Read access for `"$user`" account to `"$url`""
                 }
             }
         }
-        catch
-        {
+        catch {
             Write-LogException -Message $_
         }
-    }
-}
-#endregion
-
-#region Internet Explorer Configuration
-# ===================================================================================
-# Name: 		Disable-IEESC
-# Description:	Disable Internet Explorer Enhanced Security Configuration for administrators
-# ===================================================================================
-function Disable-IEESC
-{
-    if ($dataConfig.Configuration.Settings.DisableIEESC -eq $true)
-    {
-        Write-LogContent -Message '--------------------------------------------------------------'
-        try
-        {
-            $AdminKey = 'HKLM:\SOFTWARE\Microsoft\Active Setup\Installed Components\{A509B1A7-37EF-4b3f-8CFC-4F3A74704073}'
-            $AdminKeyValue = Get-ItemProperty -Path $AdminKey
-            if (-not ($AdminKeyValue.IsInstalled -eq '0'))
-            {
-                Write-LogContent -Message 'Disabling Internet Explorer Enhanced Security Configuration ...'
-                Set-ItemProperty -Path $AdminKey -Name 'IsInstalled' -Value 0
-            }
-            else
-            {
-                Write-LogContent -Message 'Internet Explorer ESC already Disabled - skipping.'
-            }
-        }
-        catch
-        {
-            Write-LogContent -Message 'Failed to Disable Internet Explorer Enhanced Security Configuration'
-        }
-    }
-}
-
-# ===================================================================================
-# Func: Disable-IEFirstRun
-# Desc: Disable First Run for Internet Explorer
-# ===================================================================================
-function Disable-IEFirstRun
-{
-    Write-LogContent -Message '--------------------------------------------------------------'
-    $lsaPath = 'HKCU:\Software\Microsoft\Internet Explorer\Main'
-    $lsaPathValue = Get-ItemProperty -path $lsaPath
-
-    if (-not ($lsaPathValue.DisableFirstRunCustomize -eq '1'))
-    {
-        Write-LogContent -Message 'Disabling Internet Explorer First Run ...'
-        New-ItemProperty -Path $lsaPath -Name DisableFirstRunCustomize -value '1' -PropertyType dword -Force | Out-Null
-    }
-    else
-    {
-        Write-LogContent -Message 'Internet Explorer First Run already Disabled - skipping.'
     }
 }
 #endregion
@@ -1257,179 +920,150 @@ if ($PSVersionTable.PSVersion -gt [Version]'2.0' -and $spsVersion -lt 15)
   exit
 }
 
-Write-LogContent -Message '-------------------------------------'
-Write-LogContent -Message "| Automated Script - SPSWakeUp v$spsWakeupVersion"
-Write-LogContent -Message "| Started on : $DateStarted by $currentUser"
-Write-LogContent -Message "| PowerShell Version: $psVersion"
-Write-LogContent -Message "| SharePoint Version: $spsVersion"
-Write-LogContent -Message '-------------------------------------'
+Write-Output '-------------------------------------'
+Write-Output "| Automated Script - SPSWakeUp v$spsWakeupVersion"
+Write-Output "| Started on : $DateStarted by $currentUser"
+Write-Output "| PowerShell Version: $psVersion"
+Write-Output "| SharePoint Version: $spsVersion"
+Write-Output '-------------------------------------'
 
 # Check Permission Level
-if (-NOT ([Security.Principal.WindowsPrincipal] [Security.Principal.WindowsIdentity]::GetCurrent()).IsInRole([Security.Principal.WindowsBuiltInRole] 'Administrator'))
-{
+if (-NOT ([Security.Principal.WindowsPrincipal] [Security.Principal.WindowsIdentity]::GetCurrent()).IsInRole([Security.Principal.WindowsBuiltInRole] 'Administrator')) {
     Write-Warning -Message 'You do not have Administrator rights to run this script!`nPlease re-run this script as an Administrator!'
     Break
 }
-else
-{
-    if ($Uninstall)
-    {
+else {
+    if ($Uninstall) {
         # Remove SPSWakeup script from scheduled Task
         Remove-SPSTask
     }
-    elseif ($Install)
-    {
+    elseif ($Install) {
         # Add SPSWakeup script in a new scheduled Task
         Add-SPSTask -Path $scriptRootPath
-
-        # Disable Internet Explorer Enhanced Security Configuration and First Run
-        Disable-IEESC
-        Disable-IEFirstRun
 
         # Load SharePoint Powershell Snapin
         Add-PSSharePoint
 
         # Get All Web Applications Urls
-        Write-LogContent -Message '--------------------------------------------------------------'
-        Write-LogContent -Message 'Get URLs of All Web Applications ...'
+        Write-Output '--------------------------------------------------------------'
+        Write-Output 'Get URLs of All Web Applications ...'
         $getSPWebApps = Get-SPSWebAppUrl
 
         # Add read access for Warmup User account in User Policies settings
         Add-SPSUserPolicy -Urls $getSPWebApps -UserName $UserName
     }
-    else
-    {
-        Write-LogContent -Message "Setting power management plan to `"High Performance`"..."
+    else {
+        Write-Output "Setting power management plan to `"High Performance`"..."
         Start-Process -FilePath "$env:SystemRoot\system32\powercfg.exe" `
                       -ArgumentList '/s 8c5e7fda-e8bf-4a96-9a85-a6e23a8c635c' `
                       -NoNewWindow
 
         # Load SharePoint Powershell Snapin, Assembly and System.Web
-        Add-RASharePoint
         Add-PSSharePoint
-        Add-SystemWeb
 
         # From SharePoint 2016, check if MinRole equal to Search
-        $currentSPServer = Get-SPServer | Where-Object -FilterScript {$_.Address -eq $env:COMPUTERNAME}
-        if ($null -ne $currentSPServer -and (Get-SPFarm).buildversion.major -ge 16)
-        {
-            if ($currentSPServer.Role -eq 'Search')
-            {
+        $currentSPServer = Get-SPServer | Where-Object -FilterScript { $_.Address -eq $env:COMPUTERNAME }
+        if ($null -ne $currentSPServer -and (Get-SPFarm).buildversion.major -ge 16) {
+            if ($currentSPServer.Role -eq 'Search') {
                 Write-Warning -Message 'You run this script on server with Search MinRole'
-                Write-LogContent -Message 'Search MinRole is not supported in SPSWakeUp'
+                Write-Output 'Search MinRole is not supported in SPSWakeUp'
                 Break
             }
         }
 
         # Get All Web Applications Urls, Host Named Site Collection and Site Collections
-        Write-LogContent -Message '--------------------------------------------------------------'
-        Write-LogContent -Message 'Get URLs of All Web Applications ...'
+        Write-Output '--------------------------------------------------------------'
+        Write-Output 'Get URLs of All Web Applications ...'
         $getSPWebApps = Get-SPSWebAppUrl
 
-        Write-LogContent -Message '--------------------------------------------------------------'
-        Write-LogContent -Message 'Get URLs of All Host Named Site Collection ...'
+        Write-Output '--------------------------------------------------------------'
+        Write-Output 'Get URLs of All Host Named Site Collection ...'
         $getSPSiteColN = Get-SPSHSNCUrl
 
-        Write-LogContent -Message '--------------------------------------------------------------'
-        Write-LogContent -Message 'Get URLs of All Site Collection ...'
+        Write-Output '--------------------------------------------------------------'
+        Write-Output 'Get URLs of All Site Collection ...'
         $getSPSites = Get-SPSSitesUrl
-        if ($null -ne $getSPWebApps -and $null -ne $getSPSites)
-        {
-            if ($hostEntries)
-            {
+        if ($null -ne $getSPWebApps -and $null -ne $getSPSites) {
+            if ($hostEntries) {
                 # Remove Duplicate Entries
                 $hostEntries = $hostEntries | Get-Unique
-                
+
                 # Disable LoopBack Check
-                Write-LogContent -Message '--------------------------------------------------------------'
-                Write-LogContent -Message 'Add Urls of All Web Applications or HSNC in BackConnectionHostNames regedit key ...'
-                Disable-LoopbackCheck -hostNameList $hostEntries
+                Write-Output '--------------------------------------------------------------'
+                Write-Output 'Add Urls of All Web Applications or HSNC in BackConnectionHostNames regedit key ...'
+                Disable-LoopbackCheck
 
                 # Make backup copy of the Hosts file with today's date Add Web Application and Host Named Site Collection Urls in HOSTS system File
-                Write-LogContent -Message '--------------------------------------------------------------'
-                Write-LogContent -Message 'Add Urls of All Web Applications or HSNC in HOSTS File ...'
+                Write-Output '--------------------------------------------------------------'
+                Write-Output 'Add Urls of All Web Applications or HSNC in HOSTS File ...'
 
-                foreach ($hostEntry in $hostEntries)
-                {
+                foreach ($hostEntry in $hostEntries) {
                     $hostEntryIsPresent = Select-String -Path $hostsFile -Pattern $hostEntry
-                    if ($null -eq $hostEntryIsPresent)
-                    {
-                        $hostFileNeedsUpdate = $true
-                    }
+                    if ($null -eq $hostEntryIsPresent) { $hostFileNeedsUpdate = $true }
                 }
-                if ($hostFileNeedsUpdate)
-                {
-                    Backup-HostsFile -hostsFilePath $hostsFile -hostsBackupPath $hostsFileCopy
+                if ($hostFileNeedsUpdate) {
+                    Write-Verbose -Message "Backing up $hostsFile file to: $hostsFileCopy"
+                    Copy-Item -Path $hostsFile -Destination $hostsFileCopy -Force
                     Add-HostsEntry -hostNameList $hostEntries
                 }
-                else
-                {
-                    Write-LogContent -Message 'HOSTS File already contains Urls of All Web Applications or HSNC- skipping.'
+                else {
+                    Write-Verbose -Message 'HOSTS File already contains Urls of All Web Applications or HSNC- skipping.'
                 }
             }
 
             # Request Url with System.Net.WebClient Object for All Site Collections Urls
-            Write-LogContent -Message '--------------------------------------------------------------'
-            Write-LogContent -Message 'Opening All sites Urls with Web Request Object, Please Wait...'
+            Write-Output '--------------------------------------------------------------'
+            Write-Output 'Opening All sites Urls with Web Request Object, Please Wait...'
             $InvokeResults = Invoke-SPSWebRequest -Urls $getSPSites -throttleLimit (Get-SPSThrottleLimit)
 
             # Show the results
-            foreach ($InvokeResult in $InvokeResults)
-            {
+            foreach ($InvokeResult in $InvokeResults) {
                 $resultUrl = $InvokeResult.Url
                 $resultTime = $InvokeResult.'Time(s)'
                 $resultStatus = $InvokeResult.Status
-                Write-LogContent -Message '-----------------------------------'
-                Write-LogContent -Message "| Url    : $resultUrl"
-                Write-LogContent -Message "| Time   : $resultTime seconds"
-                Write-LogContent -Message "| Status : $resultStatus"
+                Write-Output '-----------------------------------'
+                Write-Output "| Url    : $resultUrl"
+                Write-Output "| Time   : $resultTime seconds"
+                Write-Output "| Status : $resultStatus"
             }
         }
 
-        # Clean the folder of log files
-        Clear-SPSLog -path $scriptRootPath
-
-        $DateEnded = Get-date
+        $DateEnded = Get-Date
         $totalUrls = $getSPSites.Count
         $totalDuration = ($DateEnded - $DateStarted).TotalSeconds
 
-        Write-LogContent -Message '-------------------------------------'
-        Write-LogContent -Message '| Automated Script - SPSWakeUp'
-        Write-LogContent -Message "| Started on : $DateStarted"
-        Write-LogContent -Message "| Completed on : $DateEnded"
-        Write-LogContent -Message "| SPSWakeUp waked up $totalUrls urls in $totalDuration seconds"
-        Write-LogContent -Message '--------------------------------------------------------------'
-        Write-LogContent -Message '| REPORTING: Memory Usage for each worker process (W3WP.EXE)'
-        Write-LogContent -Message '| Process Creation Date | Memory | Application Pool Name'
-        Write-LogContent -Message '--------------------------------------------------------------'
+        Write-Output '-------------------------------------'
+        Write-Output '| Automated Script - SPSWakeUp'
+        Write-Output "| Started on : $DateStarted"
+        Write-Output "| Completed on : $DateEnded"
+        Write-Output "| SPSWakeUp waked up $totalUrls urls in $totalDuration seconds"
+        Write-Output '--------------------------------------------------------------'
+        Write-Output '| REPORTING: Memory Usage for each worker process (W3WP.EXE)'
+        Write-Output '| Process Creation Date | Memory | Application Pool Name'
+        Write-Output '--------------------------------------------------------------'
 
         $w3wpProcess = Get-CimInstance Win32_Process -Filter "name = 'w3wp.exe'" | Select-Object WorkingSetSize, CommandLine, CreationDate | Sort-Object CommandLine
-        foreach($w3wpProc in $w3wpProcess)
-        {
-            $w3wpProcCmdLine = $w3wpProc.CommandLine.Replace('c:\windows\system32\inetsrv\w3wp.exe -ap "','')
+        foreach ($w3wpProc in $w3wpProcess) {
+            $w3wpProcCmdLine = $w3wpProc.CommandLine.Replace('c:\windows\system32\inetsrv\w3wp.exe -ap "', '')
             $pos = $w3wpProcCmdLine.IndexOf('"')
-            $appPoolName = $w3wpProcCmdLine.Substring(0,$pos)
+            $appPoolName = $w3wpProcCmdLine.Substring(0, $pos)
             $w3wpMemoryUsage = [Math]::Round($w3wpProc.WorkingSetSize / 1MB)
-            Write-LogContent -Message "| $($w3wpProc.CreationDate) | $($w3wpMemoryUsage) MB | $($appPoolName)"
+            Write-Output "| $($w3wpProc.CreationDate) | $($w3wpMemoryUsage) MB | $($appPoolName)"
         }
-        Write-LogContent -Message '--------------------------------------------------------------'
+        Write-Output '--------------------------------------------------------------'
 
-        Trap {Continue}
-
-        # Restore backup copy of the Hosts file with today's date
-        Restore-HostsFile -hostsFilePath $hostsFile -hostsBackupPath $hostsFileCopy
+        Trap { Continue }
 
         # Clean the copy files of system HOSTS folder
         Clear-HostsFileCopy -hostsFilePath $hostsFile
-
-        Save-LogFile $pathLogFile
-
-        # Send Email with log file in attachment - For settings see XML input file
-        $mailLogContent = "Automated Script - SPSWakeUP - Started on: $DateStarted <br>"
-        $mailLogContent += "SharePoint Server : $env:COMPUTERNAME<br>"
-        Send-SPSLog -MailAttachment $pathLogFile -MailBody $mailLogContent
+        # Clean the folder of log files
+        Clear-SPSLog -path $scriptRootPath
+        #Stop Transcript parameter is equal to True
+        if ($Transcript) {
+            $pathLogFile = Join-Path -Path $scriptRootPath -ChildPath ('SPSWakeUP_script_' + (Get-Date -Format yyyy-MM-dd_H-mm) + '.log')
+            Stop-Transcript
+        }
     }
-
     Exit
 }
 #endregion