+++ CHANGELOG.md
@@ -3,6 +3,28 @@
 The format is based on and uses the types of changes according to [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - 2023-05-04
+
+### Changed
+
+- README.md
+ - BREAKING CHANGE: Windows Management Framework 5.0 is required
+ - Remove SharePoint Server 2013 from supported versions
+- Remove azure-pipelines.yml and SPSWakeUp.psd1 files
+- SPSWakeUP.ps1
+ - BREAKING CHANGE: Remove InputFile switch parameter and update examples
+ - Add Transcript boolean parameter and update examples
+ - BREAKING CHANGE: Remove logging functions and logging variables
+ - Replace Write-LogContent function with Write-Output CmdLet
+ - BREAKING CHANGE: Remove functions :
+  - Add-RASharePoint,
+  - Add-SystemWeb, 
+  - Disable-IEESC,
+  - Disable-IEFirstRun,
+  - Backup-HostsFile,
+  - Restore-HostsFile
+ - Update Disable-LoopbackCheck and Add-HostsEntry functions
+ 
 ## [2.6.1] - 2023-01-18
 
 ### Changed